backend/pyproject.toml
[project]
name = "crypto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
